+++ sharding-orchestration/sharding-orchestration-core/src/main/java/io/shardingsphere/orchestration/internal/config/event/PropertiesChangedEvent.java
@@ -15,20 +15,21 @@
  * </p>
  */
 
-package io.shardingsphere.orchestration.internal.event.config;
+package io.shardingsphere.orchestration.internal.config.event;
 
-import io.shardingsphere.core.rule.Authentication;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
+import java.util.Properties;
+
 /**
- * Authentication changed event.
+ * Properties changed event.
  *
  * @author panjuan
  */
 @RequiredArgsConstructor
 @Getter
-public final class AuthenticationChangedEvent {
+public final class PropertiesChangedEvent {
     
-    private final Authentication authentication;
+    private final Properties props;
 }
